backend/requirements.txt
requests==2.25.0
Flask==2.0.2
gunicorn==20.1.0
Flask-Cors==3.0.10
Pillow==8.0.1
pymongo[srv]==3.12.0